requirements.txt
pudb
djangorestframework
pathlib
django
psycopg2-binary
django-tastypie
